Add pyOCD erasing
